refactor: move FooLine to testclasses
